Move byte_cost_config.h file to /core/
