+++ src/core/CMakeLists.txt
@@ -34,6 +34,7 @@ target_sources(
           "models/unlock_block.c"
   PUBLIC "address.h"
          "seed.c"
+         "byte_cost_config.h"
          "utils/iota_str.h"
          "utils/byte_buffer.h"
          "utils/slip10.h"
@@ -41,7 +42,6 @@ target_sources(
          "utils/uint256.h"
          "utils/macros.h"
          "models/inputs/utxo_input.h"
-         "models/outputs/byte_cost_config.h"
          "models/outputs/feat_blocks.h"
          "models/outputs/output_alias.h"
          "models/outputs/output_extended.h"
